Rewrote compareREAL method
Now returns an integer value.
-1 if the first value is less than the second value
1 if the first value is greater than the second value
0 if the first and second values are equal

+++ real.h
@@ -14,7 +14,7 @@ typedef struct REAL REAL;
 extern REAL *newREAL(double);
 extern double getREAL(REAL *);
 extern double setREAL(REAL*, double v);
-extern double compareREAL(void *, void *);
+extern int compareREAL(void *, void *);
 extern void displayREAL(void *, FILE *);
 extern void freeREAL(void *);
